chore(eng): retire dead compat surfaces, split session god-hook
- Delete confirmed-dead components (DebugBanner, GameHeader, GameSkeleton,
  MobileNavMenu, HistoricalContext skeleton/error) and their unused
  withErrorBoundary/withGameErrorBoundary HOC wrappers; ts-prune no longer
  flags any of them.
- Split useLocalSession's authed/anonymous branching into two dedicated
  hooks (useAuthenticatedSession, useAnonymousSession) with a shared
  sessionTypes.ts contract; useLocalSession becomes a thin selector that
  calls both (rules of hooks) and picks one, so no mutator branches on
  isAuthenticated internally anymore.
- Add a TODO-debt convention (owner + tracker link) and a security-hotfix
  scope-isolation checklist to AGENTS.md so debt and hotfix scope stop
  accumulating as untracked free-form comments.
- rg confirms no live calculateOrderScore/OrderScore/saveProgress/
  loadProgress/loadSettings compatibility surface remains in src/convex.

Agent: builder-chrondle-eng-deadcode-guardrails
Agent-Surface: Claude Code workflow
Agent-Task: chrondle-eng-deadcode-guardrails
